Fix failing unit test
This failure was caused by the fact that `keytab_lookup`
returned an error when `aes_decrypt` encountered
invalid padding bytes due to an incorrect key, rather than
checking the rest of the table.

In the future, it would be best for the AES functions to
return a dedicated error code (`EIO` perhaps?) to indicate
"invalid content".